Use sslmate go-pkcs12 lib
On newer versions of MacOS p12 files are generated using sha256
hmac. This results in an error when loading the p12 file:
pkcs12: unknown digest algorithm: 2.16.840.1.101.3.4.2.1

The golang.org/x/crypto/pkcs12 pkg is frozen and will not be updated
to support sha256 hmacs. The crypto pkg recommends using
software.sslmate.com/src/go-pkcs12 which is a drop in replacement.